fixed tests and added wildcard `*` queries
