+++ src/blueprint/blueprint.rs
@@ -71,6 +71,16 @@ impl Definition {
       Definition::UnionTypeDefinition(def) => &def.name,
     }
   }
+  pub fn fields(&self) -> Option<&Vec<FieldDefinition>> {
+    match self {
+      Definition::InterfaceTypeDefinition(_) => None,
+      Definition::ObjectTypeDefinition(f) => Some(&f.fields),
+      Definition::InputObjectTypeDefinition(_) => None,
+      Definition::ScalarTypeDefinition(_) => None,
+      Definition::EnumTypeDefinition(_) => None,
+      Definition::UnionTypeDefinition(_) => None,
+    }
+  }
 }
 
 #[derive(Clone, Debug)]
